[client]: Added useForm hook and transfered form logic.
