+++ client/src/hooks/index.ts
@@ -9,3 +9,4 @@ export * from './useDateTimePicker';
 export * from './useValidations';
 export * from './useDelayedResetError';
 export * from './useTodoTotals';
+export * from './useForm';
